+++ data/Pandora.Definitions.MicrosoftGraph.Beta/CommonTypes/Model-CredentialUserRegistrationDetailsCollectionResponse.cs
@@ -14,7 +14,7 @@ namespace Pandora.Definitions.MicrosoftGraph.Beta.CommonTypes;
 internal class CredentialUserRegistrationDetailsCollectionResponseModel
 {
     [JsonPropertyName("@odata.count")]
-    public long? ODataCount { get; set; }
+    public int? ODataCount { get; set; }
 
     [JsonPropertyName("@odata.nextLink")]
     public string? ODataNextLink { get; set; }
