+++ data/Pandora.Definitions.MicrosoftGraph.Beta/CommonTypes/Model-KeyLongValuePair.cs
@@ -20,5 +20,5 @@ internal class KeyLongValuePairModel
     public string? ODataType { get; set; }
 
     [JsonPropertyName("value")]
-    public long? Value { get; set; }
+    public int? Value { get; set; }
 }
