+++ data/Pandora.Definitions.MicrosoftGraph.Beta/CommonTypes/Model-MetricTimeSeriesDataPoint.cs
@@ -20,5 +20,5 @@ internal class MetricTimeSeriesDataPointModel
     public string? ODataType { get; set; }
 
     [JsonPropertyName("value")]
-    public long? Value { get; set; }
+    public int? Value { get; set; }
 }
