+++ data/Pandora.Definitions.MicrosoftGraph.Beta/CommonTypes/Model-PrintDocument.cs
@@ -32,7 +32,7 @@ internal class PrintDocumentModel
     public string? ODataType { get; set; }
 
     [JsonPropertyName("size")]
-    public long? Size { get; set; }
+    public int? Size { get; set; }
 
     [JsonPropertyName("uploadedDateTime")]
     public DateTime? UploadedDateTime { get; set; }
